requirements.txt
autopep8==1.5.4
click==7.1.2
cycler==0.10.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pymap3d==2.4.3
pyparsing==2.4.7
pyrtlsdr==0.2.92
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
